+++ magma/types/__init__.py
@@ -1,3 +1,3 @@
 from .bit_pattern import BitPattern
 from .valid import Valid
-from .decoupled import ReadyValid, Producer, Consumer
+from .ready_valid import ReadyValid, Producer, Consumer
