+++ magma/types/ready_valid.py
